refactor(trace): move formatCostAmount and getDisplayName to shared utils, normalize tool names in trace-spans
